notifications/notifications_api/requirements.txt
aio-pika==9.0.5
asyncpg==0.27.0
backoff==2.2.1
fastapi==0.95.0
gunicorn==20.1.0
pamqp==3.2.1
pydantic==1.10.7
python-dotenv==1.0.0
python-logstash==0.4.8
orjson==3.8.10
SQLAlchemy==2.0.9
uvicorn==0.21.1